requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests
pandas
numpy
openpyxl==3.1.0.
xlrd

azure-storage-blob

opencv-python
requests_toolbelt
pyodbc
pymssql
tabula-py
camelot-py
# pdftotext
# grequests